Consolidate BAL timing infrastructure
