+++ timing/README.md
@@ -22,6 +22,15 @@ These helpers are private to the timing targets and do not add public GTSAM
 API. Legacy `gttic_` microbenchmarks continue to use `gtsam/base/timing` and are
 not candidates for migration to this harness.
 
+BAL benchmarks additionally share `BalBenchmarkConfig` and the builders in
+`timing/internal/SfmBalBenchmark`. The configuration carries ordering and
+projection-noise choices explicitly; timing headers no longer define mutable
+globals or inject the `std` and `gtsam` namespaces. Dataset selection, initial
+values, point- and camera-batch graph variants, orderings, and LM defaults live
+in that component. The end-to-end full-system and reduced-camera PCG comparison
+is isolated in `SfmPcgBenchmark` so `timeSFMBAL.cpp` remains focused on selecting
+and reporting benchmark modes.
+
 ## RangeFactor Plaza2 Benchmark
 
 This benchmark isolates the current range-only Plaza2 incremental SLAM workload
