+++ timing/sfm_ba/CMakeLists.txt
@@ -1,3 +1,4 @@
 if(GTSAM_ENABLE_CUDA AND GTSAM_ENABLE_CUDSS)
-  gtsamAddTimingGlob("*.cpp" "" "gtsam" ${GTSAM_BUILD_TIMING_ALWAYS})
+gtsamAddTimingGlob("*.cpp" "" "gtsam;gtsam_timing_utils"
+                  ${GTSAM_BUILD_TIMING_ALWAYS})
 endif()
