+++ timing/timeSFMBALnavTcam.cpp
@@ -16,32 +16,36 @@
  * @date    July 5, 2015
  */
 
-#include "timeSFMBAL.h"
-
-#include <gtsam/slam/expressions.h>
-#include <gtsam/nonlinear/ExpressionFactor.h>
 #include <gtsam/geometry/Cal3Bundler.h>
 #include <gtsam/geometry/Point3.h>
+#include <gtsam/nonlinear/ExpressionFactor.h>
+#include <gtsam/slam/expressions.h>
 
+#include "timeSFMBAL.h"
 
 using namespace std;
 using namespace gtsam;
+using symbol_shorthand::C;
+using symbol_shorthand::K;
+using symbol_shorthand::P;
+namespace bal = gtsam::timing::bal;
 
 int main(int argc, char* argv[]) {
   // parse options and read BAL file
-  SfmData db = preamble(argc, argv);
+  bal::BalBenchmarkInput input = bal::parseSmallBenchmark(argc, argv);
+  const SfmData& db = input.data;
 
   // Build graph using conventional GeneralSFMFactor
   NonlinearFactorGraph graph;
   for (size_t j = 0; j < db.numberTracks(); j++) {
     Point3_ nav_point_(P(j));
-    for (const SfmMeasurement& m: db.tracks[j].measurements) {
+    for (const SfmMeasurement& m : db.tracks[j].measurements) {
       size_t i = m.first;
       Point2 z = m.second;
       Pose3_ navTcam_(C(i));
       Cal3Bundler_ calibration_(K(i));
       graph.addExpressionFactor(
-          gNoiseModel, z,
+          input.config.projectionNoise, z,
           uncalibrate(calibration_,
                       project(transformTo(navTcam_, nav_point_))));
     }
@@ -49,14 +53,13 @@ int main(int argc, char* argv[]) {
 
   Values initial;
   size_t i = 0, j = 0;
-  for (const SfmCamera& camera: db.cameras) {
+  for (const SfmCamera& camera : db.cameras) {
     initial.insert(C(i), camera.pose());
     initial.insert(K(i), camera.calibration());
     i += 1;
   }
-  for (const SfmTrack& track: db.tracks)
-    initial.insert(P(j++), track.p);
+  for (const SfmTrack& track : db.tracks) initial.insert(P(j++), track.p);
 
   bool separateCalibration = true;
-  return optimize(db, graph, initial, separateCalibration);
+  return bal::optimize(db, graph, initial, input.config, separateCalibration);
 }
