+++ timing/timeSFMBALsmart.cpp
@@ -16,27 +16,31 @@
  * @date    Feb 26, 2016
  */
 
-#include "timeSFMBAL.h"
-
-#include <gtsam/slam/SmartProjectionFactor.h>
 #include <gtsam/geometry/Cal3Bundler.h>
 #include <gtsam/geometry/PinholeCamera.h>
 #include <gtsam/geometry/Point3.h>
+#include <gtsam/slam/SmartProjectionFactor.h>
+
+#include "timeSFMBAL.h"
 
 using namespace std;
 using namespace gtsam;
+using symbol_shorthand::C;
+namespace bal = gtsam::timing::bal;
 
 typedef PinholeCamera<Cal3Bundler> Camera;
 typedef SmartProjectionFactor<Camera> SfmFactor;
 
 int main(int argc, char* argv[]) {
   // parse options and read BAL file
-  SfmData db = preamble(argc, argv);
+  bal::BalBenchmarkInput input = bal::parseSmallBenchmark(argc, argv);
+  const SfmData& db = input.data;
 
   // Add smart factors to graph
   NonlinearFactorGraph graph;
   for (size_t j = 0; j < db.numberTracks(); j++) {
-    auto smartFactor = std::make_shared<SfmFactor>(gNoiseModel);
+    auto smartFactor =
+        std::make_shared<SfmFactor>(input.config.projectionNoise);
     for (const SfmMeasurement& m : db.tracks[j].measurements) {
       size_t i = m.first;
       Point2 z = m.second;
@@ -47,9 +51,8 @@ int main(int argc, char* argv[]) {
 
   Values initial;
   size_t i = 0;
-  gUseSchur = false;
-  for (const SfmCamera& camera : db.cameras)
-    initial.insert(C(i++), camera);
+  input.config.useSchur = false;
+  for (const SfmCamera& camera : db.cameras) initial.insert(C(i++), camera);
 
-  return optimize(db, graph, initial);
+  return bal::optimize(db, graph, initial, input.config);
 }
